+++ AssetsFlow.Server/CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [0.9.510-s] - 2025-07-...
 - User Service
 - DemoUser Seeder
+- Factory work, only 1 method open to public now 
 
 ## [0.9.503-s] - 2025-06-24
 - Add IsAnyContentMissing prop on elementDTO